# codeaware-py/openapi/current-release.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Result_AgentRunDetail_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentRunDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[AgentRunDetail]",
        "type": "object"
      },
...
      "Result_AgentRunListVO_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentRunListVO"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[AgentRunListVO]",
        "type": "object"
      },
      "Result_AgentRunReport_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentRunReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[AgentRunReport]",
        "type": "object"
      },
      "Result_AgentRunStats_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentRunStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[AgentRunStats]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CodeAware - AI 驱动的研发效能平台 (Python 重构)",
    "title": "codeaware",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/chat/agent-runs": {
      "get": {
        "description": "Agent run 列表（分页 + 可选过滤），归属 JOIN conversations。",
        "operationId": "list_agent_runs_api_chat_agent_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "conversation_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversation Id"
            }
          },
          {
            "in": "query",
            "name": "needs_review",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Needs Review"
            }
          },
          {
            "in": "query",
            "name": "review_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(pending|accepted|rejected)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Review Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AgentRunListVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Agent Runs",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/chat/agent-runs/report": {
      "get": {
        "description": "Agent 运行报表（ADR-0017 数据分析层）：聚合 agent_runs 揭示改进方向。\n\n工具使用分布（trace tool_call）、工具错误热点（tool_result error）、状态/停止原因\n分布、closure 率、平均步数、评审漏斗、最近 7 天趋势。",
        "operationId": "get_agent_run_report_api_chat_agent_runs_report_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AgentRunReport_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent Run Report",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/chat/agent-runs/stats": {
      "get": {
        "description": "Agent run 统计条（total / 待评审 / status 分布），页面头部展示。",
        "operationId": "get_agent_run_stats_api_chat_agent_runs_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AgentRunStats_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent Run Stats",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/chat/agent-runs/{turn_id}": {
      "get": {
        "description": "单 run 完整回放：metadata + trace + context_snapshot。归属不匹配返回 404。",
        "operationId": "get_agent_run_api_chat_agent_runs__turn_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AgentRunDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent Run",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/chat/agent-runs/{turn_id}/review": {
      "post": {
        "description": "评审（失败沉淀）：accepted 需 expected_tools + category；rejected 直接拒绝。",
        "operationId": "review_agent_run_api_chat_agent_runs__turn_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AgentRunListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Agent Run",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/chat/conversations": {
      "get": {
        "operationId": "list_conversations_api_chat_conversations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_list_ConversationItem__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Conversations",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/chat/conversations/{conversation_id}": {
      "delete": {
        "operationId": "delete_conversation_api_chat_conversations__conversation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Conversation",
        "tags": [
          "Chat"
        ]
      },
      "get": {
        "description": "会话消息历史（PG 真相，按时间正序）。归属不匹配返回 404（不泄露存在性）。",
        "operationId": "get_conversation_api_chat_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_list_ChatMessageVO__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Conversation",
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/api/chat/send/stream": {
      "post": {
        "description": "流式对话：typed SSE。user 经 DI 解析；直连调用时跳过归属校验。",
        "operationId": "send_stream_api_chat_send_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "版本化 typed SSE Chat 事件流"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_CONVERSATION_NOT_FOUND"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "conversation_id 不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_TURN_IN_PROGRESS"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "同一会话已有 turn 正在执行"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_REQUEST_INVALID"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "请求字段校验失败"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_START_FAILED"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "SSE 响应建立前 Transaction A 初始化失败"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Stream",
        "tags": [
          "Chat"
        ]
      }
    },
...
  }
}